adding redux
