+++ package.json
@@ -13,9 +13,12 @@
     "aws-amplify-react-native": "^2.0.1",
     "expo": "^30.0.1",
     "native-base": "^2.8.0",
+    "prop-types": "^15.6.2",
     "react": "16.3.1",
     "react-native": "https://github.com/expo/react-native/archive/sdk-30.0.0.tar.gz",
     "react-native-elements": "^0.19.1",
-    "react-navigation": "^2.14.2"
+    "react-navigation": "^2.14.2",
+    "react-redux": "^5.0.7",
+    "redux-form": "^7.4.2"
   }
 }
