+++ src/BottomTabs.js
@@ -1,6 +1,5 @@
 import { createBottomTabNavigator } from 'react-navigation'
 
-import HomeScreen from "./HomeScreen.js";
 import SignIn from './SignIn'
 import SignUp from './SignUp'
