+++ src/SignUp.js
@@ -7,10 +7,6 @@ import {
   Body,
   Header,
   Content,
-  Right,
-  Left,
-  Title,
-  Icon,
   Form,
   Item,
   Input,
